+++ omnist/__init__.py
@@ -75,7 +75,7 @@ from .errors import (
     WriteError,
 )
 
-__version__ = "0.2.4"
+__version__ = "0.2.5"
 
 __all__ = [
     # errors